+++ OrdinaryCapsule/ModConfig.cs
@@ -9,9 +9,19 @@ using System.Text;
 public class ModConfig
 {
     /// <summary>
+    ///     Gets or sets the chance that an OrdinaryCapsule will break.
+    /// </summary>
+    public float BreakChance { get; set; } = 0;
+
+    /// <summary>
     ///     Gets or sets the default production time for supported items.
     /// </summary>
     public int DefaultProductionTime { get; set; } = 1440;
+
+    /// <summary>
+    ///     Gets or sets a value indicating whether everything can be duplicated.
+    /// </summary>
+    public bool DuplicateEverything { get; set; }
 
     /// <summary>
     ///     Gets or sets a value indicating whether to unlock the recipe automatically.
@@ -22,6 +32,9 @@ public class ModConfig
     public override string ToString()
     {
         var sb = new StringBuilder();
+        sb.AppendLine($"BreakChance: {this.DefaultProductionTime.ToString(CultureInfo.InvariantCulture)}");
+        sb.AppendLine($"DefaultProductionTime: {this.DefaultProductionTime.ToString(CultureInfo.InvariantCulture)}");
+        sb.AppendLine($"DuplicateEverything: {this.DefaultProductionTime.ToString(CultureInfo.InvariantCulture)}");
         sb.AppendLine($"UnlockAutomatically: {this.UnlockAutomatically.ToString(CultureInfo.InvariantCulture)}");
         return sb.ToString();
     }